# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ubhi Backend API
  version: 1.0.0
  description: >
    API for the Ubhi brand site. Powers the public storefront (workshops, shop,
    gallery, journal, bookings, orders, newsletter and snail-mail signups) and
    the back-office admin dashboard (auth, users, customers, settings).
    Stateless JWT auth; role + per-section permission model. Storage is
    pluggable (zero-config JSON file store by default, PostgreSQL when
    DATABASE_URL is set).
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:8090
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
    OrderItem:
      type: object
      properties:
        id:
          type: string
          nullable: true
        name:
          type: string
        price:
          type: number
        qty:
          type: integer
    CreateOrderRequest:
      type: object
      required: [customer_email]
      properties:
        customer_email:
          type: string
          format: email
        customer_name:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        ship_street:
          type: string
        ship_city:
          type: string
        ship_postcode:
          type: string
        ship_country:
          type: string
        stripe_session_id:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_ref:
          type: string
        customer_name:
          type: string
        customer_email:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        status:
          type: string
        ship_street:
          type: string
        ship_city:
          type: string
        ship_postcode:
          type: string
        ship_country:
          type: string
        stripe_session_id:
          type: string
          nullable: true
        paid:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    CreateSubscriberRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        contact:
          type: string
        plan:
          type: string
        price:
          type: number
        address:
          type: string
        date_subscribed:
          type: string
        last_sent_cycle:
          type: string
    Subscriber:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        contact:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
        plan:
          type: string
        price:
          type: number
        address:
          type: string
        date_subscribed:
          type: string
        last_sent_cycle:
          type: string
        created_at:
          type: string
          format: date-time
    UpdateSubscriberRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        contact:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
        plan:
          type: string
        price:
          type: number
        address:
          type: string
        date_subscribed:
          type: string
        last_sent_cycle:
          type: string
    CreateNewsletterRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        interest:
          type: string
    NewsletterSignup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        interest:
          type: string
        created_at:
          type: string
          format: date-time
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        blurb:
          type: string
        price:
          type: number
        stock:
          type: integer
        image:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        blurb:
          type: string
        price:
          type: number
        stock:
          type: integer
        image:
          type: string
        published:
          type: boolean
...
    Setting:
      type: object
      properties:
        key:
          type: string
        value: {}
    SettingInput:
      type: object
      required: [value]
      properties:
        value: {}
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient role or permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Create an order (public checkout; totals computed server-side)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Orders]
      summary: List orders (section 'orders')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/subscribers:
    post:
      tags: [Subscribers]
      summary: Join the snail-mail list (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriberRequest'
      responses:
        '201':
          description: Created subscriber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Subscribers]
      summary: List subscribers (section 'subscribers')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscribers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscriber'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/subscribers/{id}:
    patch:
      tags: [Subscribers]
      summary: Update a subscriber (section 'subscribers')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriberRequest'
      responses:
        '200':
          description: Updated subscriber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Subscribers]
      summary: Delete a subscriber (section 'subscribers')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
  /api/updates:
    post:
      tags: [Newsletter]
      summary: Newsletter signup (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNewsletterRequest'
      responses:
        '201':
          description: Created signup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsletterSignup'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Newsletter]
      summary: List newsletter signups (section 'subscribers')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Signups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsletterSignup'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/updates/{id}:
    delete:
      tags: [Newsletter]
      summary: Delete a newsletter signup (section 'subscribers')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
...
  /api/shop:
    get:
      tags: [Shop]
      summary: List published products (public)
      responses:
        '200':
          description: Published products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags: [Shop]
      summary: Create a product (section 'shop')
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/shop/all:
    get:
      tags: [Shop]
      summary: List all products incl. unpublished (section 'shop')
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/shop/{id}:
    patch:
      tags: [Shop]
      summary: Update a product (section 'shop')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Shop]
      summary: Delete a product (section 'shop')
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
...
  /api/settings/{key}:
    get:
      tags: [Settings]
      summary: Read a setting (owner only; public for safe allowlisted keys)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Settings]
      summary: Upsert a setting (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingInput'
      responses:
        '200':
          description: Stored setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Setting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/stripe/webhook:
    post:
      tags: [Payments]
      summary: Stripe webhook (raw body, signature verified)
      description: >
        Receives the raw request body so the Stripe signature can be verified.
        On a successful checkout event, marks the matching order paid.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Event acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'